refactor(watch): carve gc.rs out of mod.rs
src/cli/watch/gc.rs (~245 lines):
- prune_last_indexed_mtime + LAST_INDEXED_PRUNE_AGE_SECS / _SIZE_THRESHOLD
- DAEMON_PERIODIC_GC_CAP_DEFAULT + daemon_periodic_gc_cap (private to gc)
- run_daemon_startup_gc, run_daemon_periodic_gc

Test-only re-imports (RebuildOutcome/RebuildResult/LAST_INDEXED_PRUNE_*) are
gated on #[cfg(test)] so non-test builds don't carry unused-import warnings.
